backend/database/db.py
import os
from dotenv import load_dotenv
from sqlmodel import create_engine, SQLModel

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
engine = create_engine(DATABASE_URL, echo=True)


def crear_tablas():
    from models.rutina import Rutina
    from models.ejercicio import Ejercicio
    SQLModel.metadata.create_all(engine)
